+++ test/unit/schema/sqlite3.js
@@ -38,6 +38,17 @@ describe("SQLite SchemaBuilder", function() {
     expect(expected).to.eql(_.map(tableSql, 'sql'));
   });
 
+  it("alter column not supported", function() {
+    try {
+      tableSql = client.schemaBuilder().alterTable('users', function(table) {
+        table.string('email').notNull().alter();
+      }).toSQL();
+      expect(false).to.eql("Should have thrown an error");
+    } catch (err) {
+      expect(err.message).to.eql("Sqlite does not support alter column.");
+    }
+  });
+
   it("drop table", function() {
     tableSql = client.schemaBuilder().dropTable('users').toSQL();
     equal(1, tableSql.length);
@@ -143,6 +154,9 @@ describe("SQLite SchemaBuilder", function() {
     equal(1, tableSql.length);
     equal(tableSql[0].sql, 'create table "users" ("foo" varchar(255), "order_id" varchar(255), foreign key("order_id") references "orders"("id"), primary key ("foo"))');
   });
+
+  // SQLite3 doesn't support named foreign keys
+  // it("adding foreign key with specific identifier throws an error");
 
   it("adding foreign key fluently", function() {
     tableSql = client.schemaBuilder().createTable('users', function(table) {